WIP/ Create the base of CreateDocumentDialogForm
