+++ src/components/CreateDocumentWidget/CreateDocumentWidget.tsx
@@ -1,7 +1,8 @@
 import { useState, type FC } from 'react';
+
 // import { currentUserStore } from '@store/index';
 import { CreateDocumentButton } from '@components/CreateDocumentButton/CreateDocumentButton';
-import { CreateDocumentForm } from '@components/CreateDocumentForm/CreateDocumentForm';
+import { CreateDocumentDialogForm } from '@components/CreateDocumentDialogForm/CreateDocumentDialogForm';
 
 export const CreateDocumentWidget: FC = () => {
   const [isMinimized, setIsMinimized] = useState<boolean>(true);
@@ -14,16 +15,16 @@ export const CreateDocumentWidget: FC = () => {
       />
     );
   return (
-    <CreateDocumentForm
+    <CreateDocumentDialogForm
+      onCancel={() => {
+        setIsMinimized(true);
+      }}
       onSubmit={(attributeValues) => {
         console.log(
           'Тут будет вызываться метод сохранения документа',
           attributeValues,
         );
         setIsMinimized(true); // После ответа от сервера
-      }}
-      onCancel={() => {
-        setIsMinimized(true);
       }}
     />
   );
